requirements.txt
django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1
openai==1.10.0
pytest==7.4.4
pytest-django==4.7.0
black==23.12.1
flake8==7.0.0
